refactor : 전체 패키지를 webserver와 application으로 분리
