+++ src/main/java/webserver/enums/RequestMethod.java
@@ -1,4 +1,4 @@
-package enums;
+package webserver.enums;
 
 public enum RequestMethod {
     GET, POST
