+++ src/main/java/webserver/exceptions/InvalidQueryParameterException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package webserver.exceptions;
 
 public class InvalidQueryParameterException extends RuntimeException {
