+++ src/main/java/webserver/exceptions/InvalidRequestException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package webserver.exceptions;
 
 public class InvalidRequestException extends RuntimeException {
     public InvalidRequestException() {
