+++ src/test/java/utils/FileIoUtilsTest.java
@@ -1,11 +1,11 @@
 package utils;
 
-import enums.ContentType;
-import org.assertj.core.api.Assertions;
+import webserver.enums.ContentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.utils.FileIoUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
