+++ src/test/java/utils/IOUtilsTest.java
@@ -4,10 +4,10 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.utils.IOUtils;
 
 import java.io.BufferedReader;
 import java.io.StringReader;
-import java.net.URLDecoder;
 import java.util.Map;
 
 public class IOUtilsTest {
